add support for GoToDesiredPosIfOtherDoesntCollide and implemented basic pig movement

+++ src/geo2/map_obj/unit_type1/player_type1.h
@@ -9,8 +9,8 @@ struct PlayerRunSpecialArgs final
 {
     double tick_len;
 
-    int mouse_x;
-    int mouse_y;
+    float mouse_x;
+    float mouse_y;
 
     bool left_pressed;
     bool right_pressed;
